TargetStrafe strafe problems fixed

+++ src/main/java/net/ccbluex/liquidbounce/utils/MovementUtils.kt
@@ -232,7 +232,7 @@ object MovementUtils : MinecraftInstance() {
 
     fun setSpeed(
             moveEvent: MoveEvent,
-            moveSpeed: Int,
+            moveSpeed: Double,
             pseudoYaw: Float,
             pseudoStrafe: Double,
             pseudoForward: Double
